Moved qualifiers to other package.
